+++ generator/page_generators/documentation/panther/documentation.js
@@ -25,6 +25,7 @@ page.end_info();
 
 
 page.text(html.link("Value Categories", "/site/documentation/panther/value_categories.html"));
+page.text(html.link("Undefined Behavior", "/site/documentation/panther/undefined_behavior.html"));